+++ web/src/main/java/org/one/gene/repository/UserRepository.java
@@ -7,6 +7,6 @@ import org.one.gene.domain.entity.User;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserRepository extends PagingAndSortingRepository<User, Integer> {
+public interface UserRepository extends PagingAndSortingRepository<User, Long> {
 }
